+++ src/application/Randomizer.h
@@ -72,8 +72,17 @@ public:
     void UpdateLevelEntrances(LevelEntrances levelEntrances);
     void UpdateMissionBlacklist(const std::vector<int>& missionBlacklist);
     void SetEntranceRandomizer(bool enableEntranceRandomizer);
-    void SetCharacterVoiceReactions(bool eggmanCommentOnCharacterUnlock, bool currentCharacterCommentOnCharacterUnlock, bool unlockedCharacterCommentOnCharacterUnlock, bool eggmanCommentOnKeyItems, bool tikalCommentOnKeyItems,
-                                   bool currentCharacterCommentOnKeyItems);
+    void SetCharacterVoiceReactions(bool eggmanCommentOnCharacterUnlock, bool currentCharacterCommentOnCharacterUnlock,
+                                    bool unlockedCharacterCommentOnCharacterUnlock, bool eggmanCommentOnKeyItems,
+                                    bool tikalCommentOnKeyItems,
+                                    bool currentCharacterCommentOnKeyItems);
+    void SetReverseControlTrapDuration(int reverseControlTrapDuration);
+    void SetTrapsOnAdventureFields(bool trapsOnAdventureFields);
+    void SetTrapsOnBossFights(bool trapsOnBossFights);
+    void SetTrapsOnPerfectChaosFight(bool trapsOnPerfectChaosFight);
+    void SetSuperSonicModRunning(bool isModRunning);
+    void SetSendDeathLinkChance(int sendDeathLinkChance);
+    void SetReceiveDeathLinkChance(int receiveDeathLinkChance);
 
 private:
     bool AreLastStoryRequirementsCompleted() const;
@@ -87,11 +96,14 @@ private:
     Options _options;
     std::string _pendingDeathCause;
     bool _deathPending;
-    
+
     float _deathLinkCooldown = 5.0f;
     std::clock_t _deathLinkCooldownTimer = -1;
     
-    
+    int _sendDeathLinkChance = 100;
+    int _receiveDeathLinkChance = 100;
+
+
     bool _eggmanCommentOnCharacterUnlock = true;
     bool _currentCharacterCommentOnCharacterUnlock = true;
     bool _unlockedCharacterCommentOnCharacterUnlock = true;
@@ -99,4 +111,5 @@ private:
     bool _eggmanCommentOnKeyItems = true;
     bool _tikalCommentOnKeyItems = true;
     bool _currentCharacterCommentOnKeyItems = true;
+    bool _superSonicModRunning = false;
 };