+++ src/input/archipelago/ArchipelagoManager.cpp
@@ -278,6 +278,7 @@ void SADX_LevelEntranceMap(const std::map<int, int> levelEntrancesValues)
     }
     randomizerPtr->UpdateLevelEntrances(levelEntrances);
 }
+
 void SADX_SetEntranceRandomizer(const int enableEntranceRandomizer)
 {
     randomizerPtr->SetEntranceRandomizer(enableEntranceRandomizer);
@@ -287,6 +288,16 @@ void SADX_SetDeathLink(const int deathLinkActive)
 {
     randomizerPtr->SetDeathLink(deathLinkActive);
 }
+void SADX_SendDeathLinkChance(const int sendDeathLinkChance)
+{
+    randomizerPtr->SetSendDeathLinkChance(sendDeathLinkChance);
+}
+
+void SADX_ReceiveDeathLinkChance(const int receiveDeathLinkChance)
+{
+    randomizerPtr->SetReceiveDeathLinkChance(receiveDeathLinkChance);
+}
+
 
 void SADX_SetRingLink(const int ringLinkActive)
 {
@@ -393,6 +404,28 @@ void SADX_BigActionStageMissions(const int missions)
     randomizerPtr->SetActionStageMissions(Characters_Big, missions);
 }
 
+
+void SADX_ReverseControlTrapDuration(const int reverseControlTrapDuration)
+{
+    randomizerPtr->SetReverseControlTrapDuration(reverseControlTrapDuration);
+}
+
+void SADX_TrapsOnAdventureFields(const int trapsOnAdventureFields)
+{
+    randomizerPtr->SetTrapsOnAdventureFields(trapsOnAdventureFields);
+}
+
+void SADX_TrapsOnBossFights(const int trapsOnBossFights)
+{
+    randomizerPtr->SetTrapsOnBossFights(trapsOnBossFights);
+}
+
+void SADX_TrapsOnPerfectChaosFight(const int trapsOnPerfectChaosFight)
+{
+    randomizerPtr->SetTrapsOnPerfectChaosFight(trapsOnPerfectChaosFight);
+}
+
+
 void ArchipelagoManager::Connect()
 {
     AP_Init(_serverIP.c_str(), "Sonic Adventure DX", playerName.c_str(), _serverPassword.c_str());
@@ -433,6 +466,9 @@ void ArchipelagoManager::Connect()
     AP_RegisterSlotDataIntCallback("EntranceRandomizer", &SADX_SetEntranceRandomizer);
 
     AP_RegisterSlotDataIntCallback("DeathLink", &SADX_SetDeathLink);
+    AP_RegisterSlotDataIntCallback("SendDeathLinkChance", &SADX_SendDeathLinkChance);
+    AP_RegisterSlotDataIntCallback("ReceiveDeathLinkChance", &SADX_ReceiveDeathLinkChance);
+    
     AP_RegisterSlotDataIntCallback("RingLink", &SADX_SetRingLink);
     AP_RegisterSlotDataIntCallback("CasinopolisRingLink", &SADX_SetCasinopolisRingLink);
     AP_RegisterSlotDataIntCallback("HardRingLink", &SADX_SetHardRingLink);
@@ -459,6 +495,14 @@ void ArchipelagoManager::Connect()
     AP_RegisterSlotDataIntCallback("AmyActionStageMissions", &SADX_AmyActionStageMissions);
     AP_RegisterSlotDataIntCallback("GammaActionStageMissions", &SADX_GammaActionStageMissions);
     AP_RegisterSlotDataIntCallback("BigActionStageMissions", &SADX_BigActionStageMissions);
+
+
+    AP_RegisterSlotDataIntCallback("ReverseControlTrapDuration", &SADX_ReverseControlTrapDuration);
+    AP_RegisterSlotDataIntCallback("TrapsOnAdventureFields", &SADX_TrapsOnAdventureFields);
+    AP_RegisterSlotDataIntCallback("TrapsOnBossFights", &SADX_TrapsOnBossFights);
+    AP_RegisterSlotDataIntCallback("TrapsOnPerfectChaosFight", &SADX_TrapsOnPerfectChaosFight);
+
+
     AP_Start();
 
     _connectedAt = std::clock();