+++ src/output/worldStateManager/WorldStateManager.cpp
@@ -916,7 +916,7 @@ FunctionHook<void, task*> onSetStartPosReturnToField(0x414500, [](task* tp)-> vo
         break;
 
     case LevelIDs_LostWorld:
-        if (CurrentCharacter == Characters_Knuckles)
+        if (CurrentCharacter == Characters_Knuckles || CurrentCharacter == Characters_Gamma)
         {
             FieldStartPos->Position = {-515.90002, 16.6, -1446.0};
             FieldStartPos->YRot = 0xC000;
@@ -1289,6 +1289,8 @@ FunctionHook<BOOL> isLostWorldBackEntranceOpen(0x53B6C0, []()-> BOOL
         return EventFlagArray[FLAG_KNUCKLES_MR_REDCUBE] && EventFlagArray[
                 FLAG_KNUCKLES_MR_BLUECUBE]
             && worldStateManagerPtr->levelEntrances.canEnter(LostWorld, CurrentCharacter);
+    if (CurrentCharacter == Characters_Gamma)
+        return worldStateManagerPtr->levelEntrances.canEnter(LostWorld, CurrentCharacter);
 
     return false;
 });
@@ -1296,7 +1298,7 @@ FunctionHook<BOOL> isLostWorldBackEntranceOpen(0x53B6C0, []()-> BOOL
 //Allows everyone to enter Lost World
 FunctionHook<BOOL> isLostWorldFrontEntranceOpen(0x532E60, []()-> BOOL
 {
-    if (CurrentCharacter == Characters_Knuckles)
+    if (CurrentCharacter == Characters_Knuckles || CurrentCharacter == Characters_Gamma)
         return false;
     return worldStateManagerPtr->levelEntrances.canEnter(LostWorld, CurrentCharacter);
 });