+++ src/pch.h
@@ -9,7 +9,7 @@
 
 #define SADX_AP_VERSION_MAJOR 0
 #define SADX_AP_VERSION_MINOR 8
-#define SADX_AP_VERSION_PATCH 7
+#define SADX_AP_VERSION_PATCH 8
 
 #define MISSION_C  2
 #define MISSION_B  1